% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_types.R
\docType{data}
\name{common_file_types}
\alias{common_file_types}
\title{Common file types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 177 rows and 4 columns.
}
\usage{
common_file_types
}
\description{
Common file types
}
\keyword{datasets}
