% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download multiple file from a url}
\usage{
download(url, ext = ask(), dest = ".", verbose = TRUE, ...)
}
\arguments{
\item{url}{the url.}

\item{ext}{ask() or file_type()}

\item{dest}{a character string (or vector, see url) with the name where the
downloaded file is saved. Tilde-expansion is performed.}

\item{verbose}{whether to print the url of the data.}

\item{...}{further options.}
}
\description{
Download multiple file from a url
}
