% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{file_type}
\alias{file_type}
\title{Helper function to split the file extensions into type}
\usage{
file_type(type)
}
\arguments{
\item{type}{category of file extension}
}
\description{
Helper function to split the file extensions into type
}
