% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{list_content}
\alias{list_content}
\title{List all the contents of a webpage}
\usage{
list_content(
  url,
  pattern = NULL,
  full = FALSE,
  negate_expr = c("[^#|^#0-9^/||^/A-Za-z+$]")
)
}
\arguments{
\item{url}{the url.}

\item{pattern}{the extension.}

\item{full}{whether to display the relative path of the absolute path.}

\item{negate_expr}{exclude some pages that do not have usefull content.}
}
\description{
List all the contents of a webpage
}
